Move program model utilities into core
