+++ crates/trust-runtime-core/src/program_model/mod.rs
@@ -1,0 +1,7 @@
+//! Portable runtime program-model helpers.
+
+#![allow(missing_docs)]
+
+pub mod util;
+
+pub use util::{method_static_storage_owner, property_setter_method_name, static_storage_name};
