+++ docs/internal/testing/checklists/architecture-improvements.md
@@ -51,6 +51,7 @@ Issue #51 init benchmark baseline evidence (2026-04-28):
 - [x] `ARCH-RTCORE-15` Move string element and IEC string helper semantics into `trust-runtime-core`, preserve the existing crate-internal `trust-runtime::value` compatibility surface for stdlib/eval/helper-eval/reference paths, and validate the slice with focused core string tests, no-default core check, runtime host compile checks, clippy, full-map doctor, and diagram drift.
 - [x] `ARCH-RTCORE-16` Move HIR-backed `SIZEOF` value/type sizing into the optional HIR-feature slice of `trust-runtime-core`, preserve the existing `trust-runtime::value` compatibility API, keep pure core no-default builds free of HIR sizing, and validate runtime `SIZEOF`, bytecode, eval, clippy, full-map doctor, and diagram drift gates.
 - [x] `ARCH-RTCORE-17` Move HIR-backed runtime default-value construction into the optional HIR-feature slice of `trust-runtime-core`, preserve the existing `trust-runtime::value` compatibility API, keep pure core no-default builds free of HIR defaults, and validate core defaults, runtime value defaults, reference defaults, clippy, full-map doctor, and diagram drift gates.
+- [x] `ARCH-RTCORE-18` Start the program-model record movement with the pure static-name helper boundary in `trust-runtime-core`, preserve the existing `trust-runtime::program_model` compatibility API, and validate core utility tests, static-local runtime integration, clippy, full-map doctor, and diagram drift gates.
 - [x] `ARCH-FULL-01` Add `full-architecture-refactor-program-checklist.md` as the umbrella board that states the runtime-core split is one branch of the full SOLID/KISS and zero-silent-bug program, not the whole program.
 - [x] `ARCH-FULL-02` Add `architecture-doctor-full-map-execution-checklist.md` so `architecture-doctor --full-map` becomes a concrete prerequisite for enforcing workspace, runtime-core, CLI, host-surface, KISS, API, and diagram rules.
 - [x] `ARCH-FULL-03` Add `hir-mutation-hardening-execution-checklist.md` for the F2 silent-bug gap: `symbol_import`, `type_check::const_eval`, and aggregate initializer validation mutation survivors.
